requirements.txt
streamlit
pandas
plotly
python-dotenv
influxdb-client




matplotlib
lightgbm
